Lots of small changes to lots of files: trying to get Python 2.2 to work again.  Some tests still fail.  Not sure if it's worth the effort...
Also contains a new checkgraph() that checks various invariants of flow
graphs, e.g. that variables are used consistently and never in more than one
block.  Graphs now have a method show() to view them using the pygame code,
which is handy to figure out what's really wrong with a graph when an
assertion in checkgraph() fails.

+++ pypy/objspace/std/test/test_intobject.py
@@ -294,7 +294,7 @@ class AppIntTest(testit.AppTestCase):
 
     def test_int_string(self):
         self.assertEquals(42, int("42"))
-        self.assertEquals(10000000000, int("10000000000"))
+        self.assertEquals(10000000000, long("10000000000"))
 
     def test_int_float(self):
         self.assertEquals(4, int(4.2))
